backend/scrobble_vault/requirements.txt
fastapi[standard]
aiohttp
aiocron
aiosqlite
python-dotenv
